@@ -20,30 +20,32 @@
   },
   "dependencies": {
     "clipboard-copy": "^4.0.1",
-    "core-js": "^3.19.1",
+    "core-js": "^3.21.0",
     "deepmerge": "^4.2.2",
+    "feather-icons": "^4.28.0",
     "leaflet": "^1.7.1",
     "leaflet.heat": "^0.2.0",
     "moment": "^2.29.1",
-    "vue": "^2.6.14",
-    "vue-ctk-date-time-picker": "^2.5.0",
-    "vue-feather-icons": "^5.1.0",
-    "vue-i18n": "^8.26.7",
-    "vue-js-modal": "^2.0.1",
-    "vue-mq": "^1.0.1",
+    "vue": "^3.2.30",
+    "vue-ctk-date-time-picker3": "^2.5.2",
+    "vue-feather": "^2.0.0",
+    "vue-i18n": "^9.2.0-beta.30",
+    "vue-final-modal": "^3.4.3",
+    "vue3-mq": "^3.0.2",
     "vue-outside-events": "^1.1.3",
-    "vue-router": "^3.5.3",
-    "vue2-leaflet": "^2.7.1",
-    "vuex": "^3.6.2"
+    "vue-router": "^4.0.12",
+    "@vue-leaflet/vue-leaflet": "^0.6.1",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.15",
-    "@vue/cli-plugin-eslint": "~4.5.15",
-    "@vue/cli-plugin-unit-jest": "~4.5.15",
-    "@vue/cli-service": "~4.5.15",
+    "@vue/cli-plugin-babel": "^4.5.15",
+    "@vue/cli-plugin-eslint": "^4.5.15",
+    "@vue/cli-plugin-unit-jest": "^4.5.15",
+    "@vue/cli-service": "^4.5.15",
+    "@vue/compiler-sfc": "^3.2.30",
     "@vue/eslint-config-prettier": "^6.0.0",
-    "@vue/test-utils": "1.3.0",
-    "babel-core": "7.0.0-bridge.0",
+    "@vue/test-utils": "^1.3.0",
+    "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^27.3.1",
     "cors-anywhere": "^0.4.4",
@@ -56,8 +58,8 @@
     "prettier": "^2.5.0",
     "sass": "^1.43.5",
     "sass-loader": "^10.1.1",
-    "vue-cli-plugin-i18n": "~2.3.1",
-    "vue-template-compiler": "^2.6.14"
+    "vue-cli-plugin-i18n": "^2.3.1",
+    "@intlify/vue-i18n-loader": "^4.1.0"
   },
   "license": "MIT",
   "repository": {
