@@ -1,9 +1,5 @@
-import Vue from "vue";
-import VueI18n from "vue-i18n";
-
+import { createI18n } from "vue-i18n";
 import config from "@/config";
-
-Vue.use(VueI18n);
 
 const locales = require.context("./locales", true, /[A-Za-z0-9-_,\s]+\.json$/i);
 const messages = {};
@@ -11,11 +7,11 @@ locales.keys().forEach((key) => {
   const matched = key.match(/([A-Za-z0-9-_]+)\./i);
   if (matched && matched.length > 1) {
     const locale = matched[1];
-    messages[locale] = locales(key);
+    messages[locale] = locales(key).default;
   }
 });
 
-export default new VueI18n({
+export default createI18n({
   locale: config.locale,
   fallbackLocale: "en-US",
   formatFallbackMessages: true,
