@@ -1,11 +1,8 @@
-import Vue from "vue";
-import Router from "vue-router";
+import { createRouter, createWebHistory } from "vue-router";
 import config from "@/config";
 import Map from "@/views/Map.vue";
 
-Vue.use(Router);
-
-export default new Router({
+export default new createRouter({
   mode: "history",
   base: config.router.basePath,
   routes: [
@@ -15,4 +12,5 @@ export default new Router({
       component: Map,
     },
   ],
+  history: createWebHistory(),
 });
